Fixed bug in cedar.h's clear() function.
The function contains a number of `if` guards that ensure that the memory allocations
only get freed at most once, by checking wether they are 0 and setting them to 0 afterwards.

However, the `if` did not use a `{}` block to guard the two statements, which lead to the
pointer always being set to 0 afterwards.

Due to the `_no_delete` flag, this can lead to a memory leak of the `_array` allocation since it
will not be freed, but its pointer forgotten.